scripts/requirements.txt
google-api-python-client==2.147.0
google-auth==2.34.0
requests==2.32.3
